+++ tests/xharness/Utilities/TempDirectory.cs
@@ -2,19 +2,19 @@
 using System.IO;
 using System.Runtime.InteropServices;
 
-namespace Xharness
+namespace Xharness.Utilities
 {
 	// A class that creates temporary directories next to the test assembly, and cleans the output on startup
 	// Advantages:
 	// * The temporary directories are automatically cleaned on Wrench (unlike /tmp, which isn't)
 	// * The temporary directories stay after a test is run (until a new test run is started),
 	//   which makes it easier to re-run (copy-paste) commands that failed.
-	public static class Cache // Not really a cache (since the root directory is cleaned in the cctor), but I couldn't come up with a better name.
+	public static class TempDirectory
 	{
 		static string root;
 		static int lastNumber;
 
-		static Cache ()
+		static TempDirectory ()
 		{
 			root = Path.Combine (Path.GetDirectoryName (System.Reflection.Assembly.GetExecutingAssembly ().Location), "tmp-test-dir");
 			if (Directory.Exists (root))
